Get visits working, with revised node package structure
This is a monster commit, I know. Sorry!

This change does a couple of things:
- Reverts to the nested `entities` and `services` .proto dirs
- Creates a build script to create the node.js buda package
- Implements visits and signup events

The reason for reverting to the nested `entities` and `services` .proto
dirs structure to get Python and Node js to play together.

For the Python package, it's best to keep the nested folder to seperate
out the modules, but the node.js protobuf package prefers it flat.

As a workaround, I extended the `budapest` tool to output a single
`buda.proto` file from the nested files, which can then be loaded
directly in the node package.

I also created a `make package-node` command that will wrap all this
logic and generate a package in the the `packages/node` directory.

To test all of this out, I've used the new Python and Node packages to
implement visit tracking on the events-collector and the buffermetrics
test server